+++ src/fragments/map-view/components/map-right-click/i18n/map-right-click.i18n.pt-br.js
@@ -7,6 +7,7 @@ export default {
     nonStopDirectionsToHere: 'Direto para aqui',
     addRouteStop: 'Adicionar parada aqui',
     centerHere: 'Centralizar aqui',
-    addAsIsochroneCenter: 'Adic. como centro de isócrono'
+    addAsIsochroneCenter: 'Adic. como centro de isócrono',
+    inspectDataOnOSM: 'Ver dados no OSM'
   }
 }